--- null-vm/.gitignore
@@ -1,2 +1,0 @@
-/target
-/program_artifacts
