+++ prover/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "riscv-prover"
+name = "prover"
 version = "0.1.0"
 edition = "2024"
