+++ vm/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "null-vm"
+name = "vm"
 version = "0.1.0"
 edition = "2024"
