+++ vm/program_artifacts/rust/basic_rust.elf
