+++ vm/programs/rust/basic_rust/Cargo.toml
