+++ vm/programs/rust/basic_rust/src/main.rs
