+++ vm/src/elf.rs
